Eliminate a large set of gcc9 warnings
I think these are relatively new warnings in gcc 8 or 9, but it's not bad to improve our style based on them.

+++ pulserecord.h
@@ -17,8 +17,8 @@ public:
     int channum;
     int presamples;
     int wordsize;
-    float sampletime;
-    float voltsperarb;
+    double sampletime;
+    double voltsperarb;
     int nsamples;
     uint64_t time_nsec;
     uint64_t serialnumber;
